running commands as script files on the containers
Ran into so many issues with sudo and tty using the default
implementation of just running lxc exec ... sudo -E #{command} that this
seems like the best way to avoid those issues. Obvious issues with this
method is that it relies on bash being there for the `bash -s < `
redirect so should probably add a way to find out what the shell is.